// src/views/character/character.types.ts
export interface CharacterProps {
  className?: string
}

export interface CharacterData {
  name: string
  title?: string
  vision: string
  weapon: string
  nation?: string
  rarity?: number
  constellation?: string
  description?: string
  affiliation?: string
  birthday?: string
  constellation_name?: string
  vision_key?: string
  weapon_type?: string
}